Rename experiment to avoid conflict
We will use some code from the experiment and share part of the implementation
