+++ client/index.js
@@ -29,7 +29,7 @@ import CapitalPage from 'capital';
 import PreviewReceiptPage from 'card-readers/preview-receipt';
 import OverviewPage from 'overview';
 import DocumentsPage from 'documents';
-import OnboardingPage from 'onboarding';
+import OnboardingExperimentPage from 'onboarding-experiment';
 import { getTasks } from 'overview/task-list/tasks';
 
 addFilter(
@@ -56,7 +56,7 @@ addFilter(
 		} );
 
 		pages.push( {
-			container: OnboardingPage,
+			container: OnboardingExperimentPage,
 			path: '/payments/onboarding',
 			wpOpenMenu: menuID,
 			breadcrumbs: [
