+++ client/onboarding-experiment/index.tsx
