+++ client/onboarding-experiment/style.scss
