+++ client/onboarding-experiment/tasks/setup-complete-task/index.tsx
@@ -8,7 +8,7 @@ import { addQueryArgs } from '@wordpress/url';
 
 import WizardTaskItem from 'additional-methods-setup/wizard/task-item';
 import WizardTaskContext from 'additional-methods-setup/wizard/task/context';
-import { OnboardingProps } from 'onboarding/types';
+import { OnboardingProps } from 'onboarding-experiment/types';
 
 interface TaskProps {
 	args: Partial< OnboardingProps >;
