+++ client/onboarding-experiment/tasks/setup-complete-task/test/index.tsx
