+++ client/onboarding-experiment/translations/descriptions.tsx
