+++ client/onboarding-experiment/translations/types.tsx
