+++ webpack.config.js
@@ -31,7 +31,6 @@ const webpackConfig = {
 		'multi-currency-analytics':
 			'./client/multi-currency-analytics/index.js',
 		order: './client/order/index.js',
-		onboarding: './client/onboarding/index.tsx',
 		'subscriptions-empty-state':
 			'./client/subscriptions-empty-state/index.js',
 		'subscription-product-onboarding-modal':
